Refactor serialization utilities.
moves opal.models.deserialize_* and opal.core.views.OpalSerializer to opal.core.serialization.

refs #1515
